pyproject.toml
[project]
name = "ansible-inventory-to-schema"
version = "0.0.0"
description = "A CLI tool to generate schema from an Ansible inventory."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["ansible", "documentation", "schema"]
authors = [
  {name = "Paul RIVIERE", email = "paul.riviere@supinfo.com"}
]

requires-python = ">=3.9"

classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Operating System :: POSIX",
  "Intended Audience :: Developers",
  "Intended Audience :: Information Technology",
  "Intended Audience :: System Administrators",
  "Topic :: Software Development :: Build Tools",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Utilities",
  "Topic :: Software Development :: Documentation",
]

dependencies = [
  "click==8.1.7",
  "ansible-core==2.16.2"
]

[project.optional-dependencies]
lint = [
  "flake8==6.0.0",
  "yamllint==1.32.0",
]
test = [
  "pytest==7.4.4",
]

[project.urls]
Documentation = "https://Paul-Riviere.github.io/ansible-inventory-to-schema/"
Issues = "https://github.com/Paul-Riviere/ansible-inventory-to-schema/issues"
Source = "https://github.com/Paul-Riviere/ansible-inventory-to-schema"

[project.scripts]
ansible-inventory-to-schema = "ansible_inventory_to_schema.cli:cli"

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"
